apps/api/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
pytest==8.3.4
httpx==0.28.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
argon2-cffi==23.1.0
python-dotenv==1.0.1
pytest-asyncio==0.25.2
